Moved ListItemProps to types file
