+++ packages/components/src/List/index.ts
@@ -30,3 +30,4 @@ export * from './ListItem'
 export * from './ListItemContext'
 export * from './ListItemLabel'
 export { listItemDimensions } from './utils'
+export { ListItemProps } from './types'
